pyproject.toml
[tool.poetry]
name = "image-template"
version = "0.1.0"
description = "Copier template to easily set up docker projects on Github"
authors = ["Tecnativa"]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^6.2.1"
pytest-xdist = "^2.2.0"
PyYAML = "^5.4.1"
plumbum = "^1.6.9"
copier = "^5.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
